+++ test_bp_vector_rmq.cpp
@@ -25,7 +25,7 @@ void test_rmq(std::vector<char> const& v, BPVector const& bitmap, std::string te
     tests.push_back(8192);
     tests.push_back(v.size());
     for (size_t t = 0; t < 10; ++t) {
-        tests.push_back(rand() % v.size());
+        tests.push_back(size_t(rand()) % v.size());
     }
     
     for(size_t t = 0; t < tests.size(); ++t) {
